Crate resolve

Domain Name System (DNS) communication protocol.

Re-exports

pub use address::address_name;

pub use config::default_config;

pub use config::DnsConfig;

pub use idna::to_ascii;

pub use idna::to_unicode;

pub use message::DecodeError;

pub use message::EncodeError;

pub use message::Message;

pub use message::Question;

pub use message::Resource;

pub use message::MESSAGE_LIMIT;

pub use record::Class;

pub use record::Record;

pub use record::RecordType;

pub use resolver::resolve_addr;

pub use resolver::resolve_host;

pub use resolver::DnsResolver;

pub use socket::DnsSocket;

pub use socket::Error;

Modules

address
IP address utility functions

config
DNS resolver configuration

hostname
System hostname detection

hosts
Implements parsing the system hosts file to produce a host table

idna
Implements RFC 3490, Internationalized Domain Names in Applications, encoding for domain name labels containing Unicode.

message
Utilities for composing, decoding, and encoding messages.

record
DNS resource record types

resolv_conf
Partial Unix resolv.conf(5) parser

resolver
High-level resolver operations

socket
Low-level UDP socket operations