Module resolve::hosts

Implements parsing the system hosts file to produce a host table

Structs

Host
Represents a single host within a host table.

HostTable
Represents a host table, consisting of addresses mapped to names.

Functions

host_file
Returns the absolute path to the system hosts file.

load_hosts
Loads a host table from the given filename.

parse_host_table
Attempts to parse a host table in the hosts file format.