Struct resolve::message::Header

pub struct Header {
    pub id: u16,
    pub qr: Qr,
    pub op: OpCode,
    pub authoritative: bool,
    pub truncated: bool,
    pub recursion_desired: bool,
    pub recursion_available: bool,
    pub rcode: RCode,
}

Represents a message header.

Fields
id: u16

Transaction ID; corresponding replies will have the same ID.

qr: Qr

Query or response

op: OpCode

Kind of query

authoritative: bool

In a response, indicates that the responding name server is an authority for the domain name in question section.

truncated: bool

Indicates whether the message was truncated due to length greater than that permitted on the transmission channel.

recursion_desired: bool

In a query, directs the name server to pursue the query recursively.

recursion_available: bool

In a response, indicates whether recursive queries are available on the name server.

rcode: RCode

Response code

Methods

impl Header

pub fn new() -> Header

Constructs a new Header with a random id value.

pub fn with_id(id: u16) -> Header

Constructs a new Header with the given id value.