Struct resolve::message::Message

pub struct Message<'a> {
    pub header: Header,
    pub question: Vec<Question>,
    pub answer: Vec<Resource<'a>>,
    pub authority: Vec<Resource<'a>>,
    pub additional: Vec<Resource<'a>>,
}

Represents a DNS message.

Fields
header: Header

Describes the content of the remainder of the message.

question: Vec<Question>

Carries the question of query type messages.

answer: Vec<Resource<'a>>

Resource records that answer the query

authority: Vec<Resource<'a>>

Resource records that point to an authoritative name server

additional: Vec<Resource<'a>>

Resource records that relate to the query, but are not strictly answers for the question.

Methods

impl<'a> Message<'a>

pub fn new() -> Message<'a>

Constructs a new Message with a random id value.

pub fn with_id(id: u16) -> Message<'a>

Constructs a new Message with the given id value.

pub fn decode(data: &[u8]) -> Result<Message, DecodeError>

Decodes a message from a series of bytes.

pub fn encode<'buf>(
&self,
buf: &'buf mut [u8]
) -> Result<&'buf [u8], EncodeError>

Encodes a message to a series of bytes. On success, returns a subslice of the given buffer containing only the encoded message bytes.

pub fn get_error(&self) -> Result<(), DnsError>

Returns a DnsError if the message response code is an error.

Important traits for RecordIter<'a>
impl<'a> Iterator for RecordIter<'a> type Item = &'a Resource<'a>;

pub fn records(&self) -> RecordIter

Returns an iterator over the records in this message.

Important traits for RecordIntoIter<'a>
impl<'a> Iterator for RecordIntoIter<'a> type Item = Resource<'a>;

pub fn into_records(self) -> RecordIntoIter<'a>

Consumes the message and returns an iterator over its records.