Struct resolve::message::MsgReader

pub struct MsgReader<'a> {
    // some fields omitted
}

Reads a single DNS message from a series of bytes.

Methods

impl<'a> MsgReader<'a>

fn new(data: &[u8]) -> MsgReader

Constructs a new message reader.

fn with_offset(data: &[u8], offset: usize) -> MsgReader

Constructs a new message reader, which will read from data, beginning at offset.

fn remaining(&self) -> usize

Returns the number of bytes remaining in the message.

fn read(&mut self, buf: &mut [u8]) -> Result<(), DecodeError>

Reads a number of bytes equal to the length of the given buffer. Returns Err(ShortMessage) if there are not enough bytes remaining.

fn read_byte(&mut self) -> Result<u8, DecodeError>

Reads a single byte from the message.

fn read_to_end(&mut self) -> Result<Vec<u8>, DecodeError>

Reads all remaining bytes.

fn read_character_string(&mut self) -> Result<Vec<u8>, DecodeError>

Read a character-string.

According to RFC 1035:

is a single length octet followed by that number of characters. is treated as binary information, and can be up to 256 characters in length (including the length octet).

fn read_u16(&mut self) -> Result<u16, DecodeError>

Reads a big-endian unsigned 16 bit integer.

fn read_u32(&mut self) -> Result<u32, DecodeError>

Reads a big-endian unsigned 32 bit integer.

fn read_into(&mut self, buf: &mut Vec<u8>, n: usize) -> Result<(), DecodeError>

Reads n bytes, which are inserted at the end of the given buffer.

fn read_name(&mut self) -> Result<String, DecodeError>

Reads a name from the message.