Struct resolve::record::Soa

pub struct Soa {
    pub mname: String,
    pub rname: String,
    pub serial: u32,
    pub refresh: u32,
    pub retry: u32,
    pub expire: u32,
    pub minimum: u32,
}

Start of authority

Fields

mname

Domain name of the name server that was the original or primary source of data for this zone.

rname

Domain name which specifies the mailbox of the person responsible for this zone.

serial

Version number of the original copy of the zone. This value wraps and should be compared using sequence space arithmetic.

refresh

Time interval before the zone should be refreshed.

retry

Time interval that should elapse before a failed refresh should be retried.

expire

Time value that specifies the upper limit on the time interval that can elapse before the zone is no longer authoritative.

minimum

Minimum TTL that should be exported with any resource record from this zone.