Enum resolve::socket::Error

pub enum Error {
    DecodeError(DecodeError),
    EncodeError(EncodeError),
    DnsError(DnsError),
    IoError(Error),
}

Represents an error in sending or receiving a DNS message.

Variants
DecodeError(DecodeError)

Error decoding received data

EncodeError(EncodeError)

Error encoding data to be sent

DnsError(DnsError)

Server responded with error message

IoError(Error)

Error generated by network operation

Methods

impl Error

pub fn is_timeout(&self) -> bool

Returns true if the error is the result of an operation having timed out.