Module resolve::message

Utilities for composing, decoding, and encoding messages.

Structs

DnsError
An error response code received in a response message.

Header
Represents a message header.

Message
Represents a DNS message.

MsgReader
Reads a single DNS message from a series of bytes.

MsgWriter
Writes a single DNS message as a series of bytes.

Question
Represents a question item.

RecordIntoIter
Yields Resource items from a Message.

RecordIter
Yields &Resource items from a Message.

Resource
Represents a resource record item.

Enums

DecodeError
Represents an error in decoding a DNS message.

EncodeError
Represents an error in encoding a DNS message.

OpCode
Represents the kind of message query.

Qr
Indicates a message is either a query or response.

RCode
Represents the response code of a message

Constants

LABEL_LIMIT
Maximum length of a name segment (i.e. a .-separated identifier).

MESSAGE_LIMIT
Maximum size of a DNS message, in bytes.

NAME_LIMIT
Maximum total length of a name, in encoded format.

Functions

generate_id
Returns a sequential ID value from a thread-local random starting value.