Struct resolve::message::Header

pub struct Header {
    pub id: u16,
    pub qr: Qr,
    pub op: OpCode,
    pub authoritative: bool,
    pub truncated: bool,
    pub recursion_desired: bool,
    pub recursion_available: bool,
    pub rcode: RCode,
}

Represents a message header.

Fields

id

Transaction ID; corresponding replies will have the same ID.

qr

Query or response

op

Kind of query

authoritative

In a response, indicates that the responding name server is an authority for the domain name in question section.

truncated

Indicates whether the message was truncated due to length greater than that permitted on the transmission channel.

recursion_desired

In a query, directs the name server to pursue the query recursively.

recursion_available

In a response, indicates whether recursive queries are available on the name server.

rcode

Response code

Methods

impl Header

fn new() -> Header

Constructs a new Header with a random id value.

fn with_id(id: u16) -> Header

Constructs a new Header with the given id value.