Struct resolve::message::MsgWriter

pub struct MsgWriter<'a> {
    // some fields omitted
}

Writes a single DNS message as a series of bytes.

Methods

impl<'a> MsgWriter<'a>

fn new(data: &mut [u8]) -> MsgWriter

Constructs a new message writer that will write into the given byte slice.

fn written(&self) -> usize

Returns the number of bytes written so far.

fn into_bytes(self) -> &'a [u8]

Returns a subslice of the wrapped byte slice that contains only the bytes written.

fn write(&mut self, data: &[u8]) -> Result<(), EncodeError>

Writes a series of bytes to the message. Returns Err(TooLong) if the whole buffer cannot be written.

fn write_character_string(&mut self, data: &[u8]) -> Result<(), EncodeError>

Write a character string, as defined by RFC 1035.

fn write_name(&mut self, name: &str) -> Result<(), EncodeError>

Writes a name to the message.

fn write_byte(&mut self, data: u8) -> Result<(), EncodeError>

Writes a single byte to the message.

fn write_u16(&mut self, data: u16) -> Result<(), EncodeError>

Writes an unsigned 16 bit integer in big-endian format.

fn write_u32(&mut self, data: u32) -> Result<(), EncodeError>

Writes an unsigned 32 bit integer in big-endian format.