Module resolve::resolv_conf

Partial Unix resolv.conf(5) parser

Constants

DEFAULT_ATTEMPTS
Default value of "options attempts:n"

DEFAULT_N_DOTS
Default value of "options ndots:n"

DEFAULT_TIMEOUT
Default value of "options timeout:n"

MAX_ATTEMPTS
Maximum allowed value of "options attempts:n"

MAX_NAME_SERVERS
Maximum number of name servers loaded from resolv.conf

MAX_N_DOTS
Maximum allowed value of "options ndots:n"

MAX_TIMEOUT
Maximum allowed value of "options timeout:n"

RESOLV_CONF_PATH
Path to system resolv.conf

Functions

load
Examines system resolv.conf and returns a configuration loosely based on its contents. If the file cannot be read or lacks required directives, an error is returned.