Enum resolve::socket::Error

pub enum Error {
    DecodeError(DecodeError),
    EncodeError(EncodeError),
    DnsError(DnsError),
    IoError(Error),
}

Represents an error in sending or receiving a DNS message.

Variants

DecodeError

Error decoding received data

EncodeError

Error encoding data to be sent

DnsError

Server responded with error message

IoError

Error generated by network operation

Methods

impl Error

fn is_timeout(&self) -> bool

Returns true if the error is the result of an operation having timed out.